Tests + WIP-solution for 2015 day 6 part 1.
Terrible performance, needs to be reimplemented.
I wanted to create a set of pure functions to solve the problem without
sharing state. However, Object.assign() on objects with more than 500.000
keys are far from performant.

Pure functions are cvery clean to read and easy to understand. But in this
case, they hurt performance too much to be viable.